Add files via upload
Titanic - Sinking of the unsinkable 
The RMS Titanic, a luxury steamship, sank in the early hours of April 15, 1912, off the coast of Newfoundland in the North Atlantic after sideswiping an iceberg during its maiden voyage. Of the 2,240 passengers and crew on board, more than 1,500 lost their lives in the disaster.

Description:¶
The titanic Dataset is analysed and with machine learning technique, survival prediction is made to find out accurate survival rate.